Actualización de cuentas de ahorro completada

+++ Simulador de banco/src/DAO/interfazCuentaAhorroDAO.java
@@ -9,6 +9,6 @@ import Modelo.CuentaDeAhorro;
 public interface interfazCuentaAhorroDAO {
     List<CuentaDeAhorro> obtenerCuentasDeAhorro();
     boolean agregarCuentaDeAhorro(CuentaDeAhorro cuentaDeAhorro) throws SQLException ;
-    boolean actualizarCuentaDeAhorro(int idCliente, int saldo, float tasaInteres, int topeMinimo) throws SQLException ;
+    boolean actualizarCuentaDeAhorro(int idCliente, int saldo, double tasaInteres, int topeMinimo) throws SQLException ;
     boolean eliminarCuentaDeAhorro(int idCliente) throws SQLException ;
 }
